% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiSpaR-package.R
\docType{package}
\name{HiSpaR-package}
\alias{HiSpaR-package}
\alias{HiSpaR}
\title{HiSpaR: Hierarchical Inference of Spatial Positions from Hi-C Data}
\description{
HiSpaR provides R bindings for the HiSpa C++ library, enabling hierarchical
Bayesian inference of 3D chromatin structures from Hi-C contact matrices.
}
\details{
The package provides functions to:
\itemize{
  \item Analyze Hi-C contact matrices using MCMC sampling
  \item Infer 3D chromatin structures
  \item Use prior information from other datasets
  \item Perform cluster-based hierarchical analysis
}

Main functions:
\itemize{
  \item \code{\link{hispa_analyze}}: Run complete HiSpa analysis
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/masterStormtrooper/HiSpa}
  \item Report bugs at \url{https://github.com/masterStormtrooper/HiSpa/issues}
}

}
\author{
\strong{Maintainer}: Your Name \email{your.email@example.com}

Other contributors:
\itemize{
  \item Contributors Names [contributor]
}

}
\keyword{internal}
